Gives insight in enrolments year after degree

Gives insight in enrolments year after degree

Usage

after_degree(program_of_interest, my_vlpbek_compact, prop_exam = FALSE)

Arguments

program_of_interest

string containing the 5-character program code of interest

my_vlpbek_compact

a vlpbek_compact object

prop_exam

boolean, default FALSE, set to TRUE for after propedeutic exam

Value

a list with 10 objects

program_of_interest

the program code that was entered

degree

D for propedeutic, A for AD, B for Bachelor or M for Master

brin_own

brin_own from the my_vlpbek_compact object

academic_year_from

the earliest academic year in the data

academic_year_to

the latest academic year in the data

n_students

number of students with a degree from brin_own/program_of_interest in academic_year_from

n_students_enrolment_type_single

number of above students with at most one enrolment brin/program per academic year

n_students_enrolment_type_multiple

difference between the two numbers of students

from_to_single

dataframe with 7 columns with counts per flow, for the students with enrolment_type = single

from_to_multiple

dataframe with 7 columns with flow information per (anonymized) student, for the students with enrolment_type = single

Examples

if (FALSE) { # \dontrun{
after_degree(program_of_interest = "34507",
 my_vlpbek_compact = my_vlp_compact_thing,
  prop_exam = TRUE)
} # }