Plots the global flows between brin_own, other HE and not in HE

plot_brinflows() takes the summary_presences_brin from a flow_insights list, and turns it into a sankey diagram

Usage

plot_brinflows(
  my_flow_insights,
  color_brin_own = "#853887",
  color_other_he = "#246ABE",
  color_brin_own_other_he = "#00B17E",
  color_outside_he = "grey",
  label_brin_own = "this HEI",
  label_other_he = "other HEI",
  label_outside_he = "outside HE",
  padding = 2,
  alpha = 0.4,
  display_labels = TRUE,
  match_label_color = TRUE,
  label_fontface = "bold",
  label_size = 10
)

Arguments

my_flow_insights

a flows_insights object

color_brin_own

string containing color code used for brin_own

color_other_he

string containing color code used for other HE

color_brin_own_other_he

string containing color code used for brin_own & other HE

color_outside_he

string containing color code used for outside HE

label_brin_own

string to use instead of "brin_own"

label_other_he

string to use instead of "other_HE"

label_outside_he

string to use instead of "outside_HE"

padding

constant used for displaying space between flows

alpha

numeric, transparancy of colors (0 = complete transparent, 1 = opaque)

display_labels

boolean used for displaying labels or not

match_label_color

boolean used to indicate labels in same color as nodes

label_fontface

string used for styling the font of labels

label_size

number indicating the font size of labels

Value

a ggplot, displaying the flows between brin_own, other HE and not in HE

Examples

if (FALSE) { # \dontrun{
plot_brinflows(my_flow_insights)
} # }